+++ lib/domain/diary/confirm_diary_store.dart
